requirements.txt
python-telegram-bot==12.8
beautifulsoup4==4.9.0
requests==2.24.0
twilio==6.44.2
html5lib==1.1
requests[socks]
requests[security]
stem==1.8.0
fake-useragent==0.1.11
